Sara branch (#16)
* added some comments

* added tinymce editor create and show pages

I added TinyMCE as WISAWIG editor in /home/editor.
When the submit button is clicked, the page is saved in the database in Page table. Page.sql file can create the table and some sample pages.
I created /home/show?title=PageTitle to show the content of previously created pages.